+++ website/app/(default)/page.tsx
@@ -334,6 +334,7 @@ LISN.watchers.ScrollWatcher.onScroll(
           <li>Wrappers for more frameworks.</li>
           <li>SmoothScroller widget.</li>
           <li>Radio/checkbox trigger.</li>
+          <li>Default pager transitions to use gesture momentum.</li>
           <li>Custom duration for ScrollTo action.</li>
           <li>Passing information on elements spanned by a gesture.</li>
           <li>