+++ website/app/(demos)/demos/pager-slider/demo.tsx
@@ -6,7 +6,7 @@ import {
   PagerPrevSwitchComponent,
   PagerNextSwitchComponent,
 } from "@lisn.js/react";
-import "lisn.js/pager.css";
+import "lisn.js/pager-slider.css";
 
 import Image from "next/image";